+++ src/main/java/model/ExchangeRate.java
@@ -1,9 +1,11 @@
 package model;
 
+import java.math.BigDecimal;
+
 public record ExchangeRate(
         int id,
         Currency baseCurrency,
         Currency targetCurrency,
-        double rate
+        BigDecimal rate
 ) {
 }
